+++ src/main/kotlin/io/snyk/plugin/settings/SnykProjectSettingsConfigurable.kt
@@ -21,7 +21,6 @@ import io.snyk.plugin.snykcode.core.SnykCodeParams
 import io.snyk.plugin.snykcode.newCodeRestApi
 import io.snyk.plugin.ui.SnykBalloonNotificationHelper
 import io.snyk.plugin.ui.SnykSettingsDialog
-import org.eclipse.lsp4j.DidChangeConfigurationParams
 import snyk.amplitude.api.ExperimentUser
 import snyk.common.lsp.LanguageServerWrapper
 import snyk.common.toSnykCodeApiUrl
@@ -99,14 +98,10 @@ class SnykProjectSettingsConfigurable(val project: Project) : SearchableConfigur
             snykProjectSettingsService?.additionalParameters = snykSettingsDialog.getAdditionalParameters()
         }
 
-        val wrapper = LanguageServerWrapper.getInstance()
-        val params = DidChangeConfigurationParams(wrapper.getSettings())
-        wrapper.languageServer.workspaceService.didChangeConfiguration(params)
-
         if (isSnykCodeLSEnabled()) {
             runBackgroundableTask("Updating Snyk Code settings", project, true) {
                 settingsStateService.isGlobalIgnoresFeatureEnabled =
-                    wrapper.getFeatureFlagStatus("snykCodeConsistentIgnores")
+                    LanguageServerWrapper.getInstance().getFeatureFlagStatus("snykCodeConsistentIgnores")
             }
         }