+++ src/main/kotlin/snyk/common/lsp/LanguageServerWrapper.kt
@@ -1,6 +1,5 @@
 package snyk.common.lsp
 
-import com.intellij.ide.impl.ProjectUtil
 import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
@@ -21,6 +20,7 @@ import org.eclipse.lsp4j.CodeActionCapabilities
 import org.eclipse.lsp4j.CodeLensCapabilities
 import org.eclipse.lsp4j.CodeLensWorkspaceCapabilities
 import org.eclipse.lsp4j.DiagnosticCapabilities
+import org.eclipse.lsp4j.DidChangeConfigurationParams
 import org.eclipse.lsp4j.DidChangeWorkspaceFoldersParams
 import org.eclipse.lsp4j.ExecuteCommandParams
 import org.eclipse.lsp4j.InitializeParams
@@ -39,13 +39,14 @@ import snyk.common.getEndpointUrl
 import snyk.common.lsp.commands.ScanDoneEvent
 import snyk.pluginInfo
 import snyk.trust.WorkspaceTrustService
+import snyk.trust.confirmScanningAndSetWorkspaceTrustedStateIfNeeded
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
 import java.util.concurrent.Future
 import java.util.concurrent.TimeUnit
+import java.util.concurrent.locks.ReentrantLock
 import kotlin.io.path.exists
 
-private const val DEFAULT_SLEEP_TIME = 100L
 private const val INITIALIZATION_TIMEOUT = 20L
 
 @Suppress("TooGenericExceptionCaught")
@@ -76,24 +77,24 @@ class LanguageServerWrapper(
      */
     lateinit var process: Process
 
-    private var isInitializing: Boolean = false
+    private var isInitializing: ReentrantLock = ReentrantLock()
 
-    val isInitialized: Boolean
+    internal val isInitialized: Boolean
         get() = ::languageClient.isInitialized &&
             ::languageServer.isInitialized &&
             ::process.isInitialized &&
             process.info().startInstant().isPresent &&
-            process.isAlive
+            process.isAlive &&
+            !isInitializing.isLocked
 
     @OptIn(DelicateCoroutinesApi::class)
-    internal fun initialize() {
+    private fun initialize() {
         if (lsPath.toNioPathOrNull()?.exists() == false) {
             val message = "Snyk Language Server not found. Please make sure the Snyk CLI is installed at $lsPath."
             logger.warn(message)
             return
         }
         try {
-            isInitializing = true
             val snykLanguageClient = SnykLanguageClient()
             languageClient = snykLanguageClient
             val logLevel = if (snykLanguageClient.logger.isDebugEnabled) "debug" else "info"
@@ -111,16 +112,15 @@ class LanguageServerWrapper(
             }
 
             launcher.startListening()
-
             sendInitializeMessage()
         } catch (e: Exception) {
             logger.warn(e)
-        } finally {
-            isInitializing = false
+            process.destroy()
         }
 
         // update feature flags
-        pluginSettings().isGlobalIgnoresFeatureEnabled = getFeatureFlagStatus("snykCodeConsistentIgnores")
+        pluginSettings().isGlobalIgnoresFeatureEnabled =
+            getFeatureFlagStatusInternal("snykCodeConsistentIgnores")
     }
 
     fun shutdown(): Future<*> {
@@ -143,6 +143,8 @@ class LanguageServerWrapper(
     }
 
     private fun getTrustedContentRoots(project: Project): MutableSet<VirtualFile> {
+        if (!confirmScanningAndSetWorkspaceTrustedStateIfNeeded(project)) return mutableSetOf()
+
         // the sort is to ensure that parent folders come first
         // e.g. /a/b should come before /a/b/c
         val contentRoots = project.getContentRootVirtualFiles().filterNotNull().sortedBy { it.path }
@@ -219,11 +221,13 @@ class LanguageServerWrapper(
     }
 
     fun ensureLanguageServerInitialized(): Boolean {
-        while (isInitializing) {
-            Thread.sleep(DEFAULT_SLEEP_TIME)
-        }
         if (!isInitialized) {
-            initialize()
+            try {
+                isInitializing.lock()
+                initialize()
+            } finally {
+                isInitializing.unlock()
+            }
         }
         return isInitialized
     }
@@ -241,13 +245,8 @@ class LanguageServerWrapper(
         }
     }
 
-    fun sendScanCommand() {
+    fun sendScanCommand(project: Project) {
         if (!ensureLanguageServerInitialized()) return
-        val project = ProjectUtil.getActiveProject()
-        if (project == null) {
-            logger.warn("No active project found, not sending scan command.")
-            return
-        }
         getTrustedContentRoots(project).forEach {
             sendFolderScanCommand(it.path)
         }
@@ -255,6 +254,10 @@ class LanguageServerWrapper(
 
     fun getFeatureFlagStatus(featureFlag: String): Boolean {
         ensureLanguageServerInitialized()
+        return getFeatureFlagStatusInternal(featureFlag)
+    }
+
+    private fun getFeatureFlagStatusInternal(featureFlag: String): Boolean {
         if (!isSnykCodeLSEnabled()) {
             return false
         }
@@ -269,17 +272,16 @@ class LanguageServerWrapper(
             val ok = resultMap?.get("ok") as? Boolean ?: false
             val userMessage = resultMap?.get("userMessage") as? String ?: "No message provided"
 
-
             if (ok) {
                 logger.info("Feature flag $featureFlag is enabled.")
                 return true
             } else {
-                logger.warn("Feature flag $featureFlag is disabled. Message: $userMessage")
+                logger.info("Feature flag $featureFlag is disabled. Message: $userMessage")
                 return false
             }
 
         } catch (e: Exception) {
-            logger.error("Error while checking feature flag: ${e.message}", e)
+            logger.warn("Error while checking feature flag: ${e.message}", e)
             return false
         }
     }
@@ -318,6 +320,21 @@ class LanguageServerWrapper(
             integrationName = pluginInfo.integrationName,
             integrationVersion = pluginInfo.integrationVersion,
         )
+    }
+
+    fun updateConfiguration() {
+        ensureLanguageServerInitialized()
+        val params = DidChangeConfigurationParams(getInstance().getSettings())
+        languageServer.workspaceService.didChangeConfiguration(params)
+        if (pluginSettings().scanOnSave) {
+            ProjectManager.getInstance().openProjects.forEach { sendScanCommand(it) }
+        }
+    }
+
+    fun addContentRoots(project: Project) {
+        if (!isInitialized) return
+        val added = getWorkspaceFolders(project)
+        updateWorkspaceFolders(added, emptySet())
     }
 
     companion object {